% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchTrackPath.R
\name{batchTrackPath}
\alias{batchTrackPath}
\title{Track an animal's movement across a series of still frames}
\usage{
batchTrackPath(
  input.dir,
  output.filename = NULL,
  fps,
  box = 1,
  liveplot = FALSE
)
}
\arguments{
\item{input.dir}{a character string specifying a directory containing only
jpeg files extracted from a video}

\item{output.filename}{a character string specifying the path of an output
track file (in CSV format). If this is NULL (default) then no file will be
written.}

\item{fps}{an integer specifying the frame rate at which jpegs were extracted
from a video; this value is used for distance and velocity calculations,
this is required for accurate timestamp calculations}

\item{box}{an integer specifying the size of the tracking box relative to the
initial animal selection box; a larger box size will help prevent the
animal being lost (useful for fast moving animals) but will also increase
sensitivity to background and lighting changes (default 1)}

\item{liveplot}{if TRUE, the image fram will be shown with a red point at the
tracked position. This makes tracking very much slower but is usseful if
tracking a new animal or using a new arena for the first time to help
adjust detection parameters. Default is FALSE.}
}
\value{
A list containing the position/co-ordinates of the animal (matrix
  with columns 'x' and 'y') and the timestamp (matrix with 1 column 't') for
  each frame, as well as some quality control metrics
}
\description{
This function is the core of the \code{pathtrackr} package. It takes a series
of jpegs (extracted from a video) as inputs and tracks an animal's movement
across frames. A list is returned containing the xy co-ordinates of the
animal in each frame, as well as summary statistics. The returned list can be
called for plotting and further functions in the \code{pathtrackr} package.
}
\details{
\code{trackPath} tracks an individual animal's movement across a
  series of still frames. The function utilises a focusing box to limit the
  search area for the animal relative to its previous position. This makes
  \code{trackPath} relatively robust to background lighting changes,
  extraneous background movement and jpeg noise. It can  handle a dark animal
  on a light background and vice versa, as well as heterogeneous backgrounds
  where the animal is at times darker and lighter than the background.
}
