% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countCells.R
\name{countCells}
\alias{countCells}
\title{Overlay a grid and count cell crossings}
\usage{
countCells(path.list, xgrid, ygrid)
}
\arguments{
\item{path.list}{a list object created by the \code{trackPath} function}

\item{xgrid}{an integer specifying the number of grid cells the arena should be divided into along its width}

\item{ygrid}{an integer specifying the number of grid cells the arena should be divided into along its height}
}
\value{
If assigned to an object, a list containing the number of cells crossed and the proportion of total cells crossed; if not assigned to an object, a message in the console is returned.
}
\description{
This function overlays a grid on the track generated by the \code{trackPath} function and counts the number of cells crossed by the animal. It is useful for returning count and proportional data abount an animal's path.
}
