% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyJPEG.R
\name{greyJPEG}
\alias{greyJPEG}
\title{Greyscale conversion}
\usage{
greyJPEG(filepath)
}
\arguments{
\item{filepath}{a charcter string specifying the location of the video file}
}
\value{
A matrix of pixel values
}
\description{
Open a jpg file and convert it to greyscale.
}
