% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVideo.R
\name{makeVideo}
\alias{makeVideo}
\title{Make a video of an animal's tracked path}
\usage{
makeVideo(dirpath, xarena, yarena, fps, box = 2, contrast = 0.5)
}
\arguments{
\item{dirpath}{a character string specifying a directory containing only jpeg files extracted from a video}

\item{xarena}{an integer specifying the arena width in mm}

\item{yarena}{an integer specifying the arena height in mm}

\item{fps}{an integer specifying the frame rate at which jpegs were extracted from a video}

\item{box}{an integer specifying the size of the tracking box relative to the initial animal selection box; a larger box size will help prevent the animal being lost (useful for fast moving animals) but will also increase sensitivity to background and lighting changes (default 2)}

\item{contrast}{a value between 0 and 1 specifying the contrast threshold for distinguishing between the animal and background; by default \code{contrast} is set to 0.5, which will suit most situations, but may need adjusting for very low contrast videos}
}
\value{
An mp4 video of an animal's movement containing the original video, tracking behaviour plots and summary plots.
}
\description{
\code{makeVideo} uses a \code{trackPath} and FFmpeg to produce an mp4 video of an animal's movement along with tracking behaviour and summary plots.
}
\details{
See documentation for \code{\link{trackPath}}.
}
\note{
\code{makeVideo} requires FFmpeg to be installed on your machine. FFmpeg is a cross-platform, open-source video editing tool. It can be downloaded from \url{https://ffmpeg.org}.
}

