% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualPath.R
\name{manualPath}
\alias{manualPath}
\title{Manually track an animal's movement across a series of still frames}
\usage{
manualPath(dirpath, xarena, yarena, fps = 30)
}
\arguments{
\item{dirpath}{a character string specifying a directory containing only jpeg files extracted from a video}

\item{xarena}{an integer specifying the arena width in mm; this value is used for distance and velocity calculations, an incorrect value will not cause an error but will result in inaccurate calculations}

\item{yarena}{an integer specifying the arena height in mm; this value is used for distance and velocity calculations, an incorrect value will not cause an error but will result in inaccurate calculations}

\item{fps}{an integer specifying the frame rate at which jpegs were extracted from a video; this value is used for distance and velocity calculations, an incorrect value will not cause an error but will result in inaccurate calculations}
}
\value{
A list containing a matrix of xy co-ordinates of the animal in each frame, a matrix of movement data including the distance, velocity and trajectories of movement between frames, and summary statistics.
}
\description{
This function is similar to the core \code{trackPath} of the \code{pathtrackr} package. However, rather than automatically tracking the animal, it requires the user to manual click on the position of the animal in each frame. This function is useful when your video is not appropriate for automatic video tracking, such as very low contrast videos, or videos with a lot of background movement. The same output as \code{trackPath} will be produced, i.e. a list containing the xy co-ordinates of the animal in each frame, as well as summary statistics. The returned list can be called for plotting and further functions in the \code{pathtrackr} package.
}
\details{
\code{manualPath} tracks an individual animal's movement across a series of still frames by requiring the user to manually click onht eanimal's position in each frame.
}
