% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathSummary.R
\name{pathSummary}
\alias{pathSummary}
\title{Summary table of an animal's tracked path}
\usage{
pathSummary(path.list)
}
\arguments{
\item{path.list}{a list object created by the \code{trackPath} function}
}
\value{
A dataframe with summary info of an animal's path.
}
\description{
This function produces a summary table of total distance, mean velocity, total duration and the number of frames where tracking was unsuccessful, based on a list object generated by the \code{trackPath} function.
}
