% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPath.R
\name{plotPath}
\alias{plotPath}
\title{Plot animal movement data}
\usage{
plotPath(path.list)
}
\arguments{
\item{path.list}{a list object created by the \code{trackPath} function}
}
\value{
A density plot of time spent in a position overlaid with the tracked animal's path.
}
\description{
This function plots the path of an animal tracked using the \code{trackPath} function.
}
\details{
This functions relies on the \code{smoothScatter} function from the base \code{graphics} package.
}

