% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummary.R
\name{plotSummary}
\alias{plotSummary}
\title{Summary plots of an animal's tracked path}
\usage{
plotSummary(path.list)
}
\arguments{
\item{path.list}{a list object created by the \code{trackPath} function}
}
\value{
A four panel summary plot.
}
\description{
This function produces a summary plot of cumulative distance and velocity over time, and rose plots of absolute and relative bearings, based on a list object generated by the \code{trackPath} function.
}
\details{
\code{plotSummary} uses \code{ggplot2} and \code{gridExtra} to produce a four panel summary plot.
}

