% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflect.R
\name{reflect}
\alias{reflect}
\title{Matrix reversal by row}
\usage{
reflect(mat)
}
\arguments{
\item{mat}{a matrix to be reflected around horizontal axis}
}
\value{
A matrix
}
\description{
Reverse row order of matrix for correctly oriented raster plots.
}
