% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitVideo.R
\name{splitVideo}
\alias{splitVideo}
\title{Compress and split a video into jpeg files}
\usage{
splitVideo(filepath, fps, xpix = 320, ypix = -1)
}
\arguments{
\item{filepath}{a charcter string specifying the location of the video file}

\item{fps}{an integer: the number of frames per second of video to keep; if \code{fps} is the as the frame rate of the video, all frames will be saved}

\item{xpix}{an integer: the width of saved jpegs in pixels; the default is 320}

\item{ypix}{an integer: the height of saved jpegs in pixels; the default is -1, which will preserve the aspect ratio}
}
\value{
Returns a new directory (in the same directory as the video file) containing each video frame as a jpeg file.
}
\description{
\code{splitVideo} uses FFmpeg via a system call to compress a video file and split it into separate jpeg files in a new directory.
}
\details{
The tracking functions within the \code{pathtrackr} package do not require high resolution imagery. It is is recommened that videos are compressed using the \code{xpix} and \code{ypix} variables to improve processing time. If FFmpeg is not available, video frames can be extracted using other video editing software.
}
\note{
\code{splitVideo} requires FFmpeg to be installed on your machine. FFmpeg is a cross-platform, open-source video editing tool. It can be downloaded from \url{https://ffmpeg.org}.
}
