% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitVideo.R
\name{splitVideo}
\alias{splitVideo}
\title{Compress and split a video into jpeg files}
\usage{
splitVideo(filepath, fps, xpix = -1, ypix = -1)
}
\arguments{
\item{filepath}{a character string specifying the location of the video file.}

\item{fps}{an integer: the number of frames per second of video to keep; if
\code{fps} is the same as the frame rate of the video, all frames will be
saved.}

\item{xpix}{an integer: the width of saved jpegs in pixels; the default of -1
will use the native width of the video without rescaling.}

\item{ypix}{an integer: the height of saved jpegs in pixels; the default is
-1, which will preserve the aspect ratio.}
}
\value{
Returns a new directory (in the same directory as the video file)
  containing each video frame as a jpeg file.
}
\description{
\code{splitVideo} uses FFmpeg via a system call to compress a video file and
split it into separate jpeg files in a new directory.
}
\details{
The tracking functions within the \code{pathtrackr} package do not
  require high resolution imagery. It is is recommended that videos are
  compressed using the \code{xpix} and \code{ypix} variables to improve
  processing time. If FFmpeg is not available, video frames can be extracted
  using other video editing software.
}
\note{
\code{splitVideo} requires FFmpeg to be installed on your machine.
  FFmpeg is a cross-platform, open-source video editing tool. It can be
  downloaded from \url{https://ffmpeg.org}. For Mac OS X, easier installation
  options exist; such as via Homebrew (\url{https://brew.sh/}) with
  \code{brew install ffmpeg}. For Linux, it is even simpler as it should
  already be in your package repository (e.g. \code{sudo apt install ffmpeg}
  for Ubuntu).
}
